rename ZePages and ZeTheme packages and zeclasses
