+++ zeapp/app/src/main/java/de/berlindroid/zeapp/ui/zetheme/ZeType.kt
@@ -1,4 +1,4 @@
-package de.berlindroid.zeapp.ui.theme
+package de.berlindroid.zeapp.ui.zetheme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
